Separate classic/universal tests
